fix(test): isolate HOME in E2E tests to prevent config interference
E2E tests were failing on macOS because they loaded the user's global
config from ~/.codi/config.json, which could override the mock provider
settings. The global config's model map routing would interfere with
the test's mock provider.

Changes:
- Add shared ProcessHarness in tests/helpers/process-harness.ts that
  automatically creates an isolated temp HOME directory
- Update all E2E test files to use the shared ProcessHarness
- Add write(), clearOutput(), and pid getter to ProcessHarness for
  tests that need them
- ProcessHarness now detects when tests provide their own HOME
  (like session-commands tests) and doesn't override it

This reduces code duplication (~650 lines removed) and ensures tests
run in isolated environments regardless of the user's global config.
