+++ tests/commands.e2e.test.ts
@@ -8,23 +8,14 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-
-// Platform-aware timeouts - macOS CI is slower
-const isMacOS = process.platform === 'darwin';
-const TEST_TIMEOUT = isMacOS ? 90000 : 20000;
-const WAIT_TIMEOUT = isMacOS ? 60000 : 15000;
-
-// Set longer timeout for E2E tests
-vi.setConfig({ testTimeout: TEST_TIMEOUT });
 import * as path from 'node:path';
 import * as fs from 'node:fs';
 import * as os from 'node:os';
-import { spawn, type ChildProcess } from 'node:child_process';
 import { setupMockE2E, cleanupMockE2E, textResponse, type MockE2ESession } from './helpers/mock-e2e.js';
+import { ProcessHarness, TEST_TIMEOUT, distEntry } from './helpers/process-harness.js';
 
-function distEntry(): string {
-  return path.resolve(process.cwd(), 'dist', 'index.js');
-}
+// Set longer timeout for E2E tests
+vi.setConfig({ testTimeout: TEST_TIMEOUT });
 
 function createTempProjectDir(): string {
   const dir = path.join(os.tmpdir(), `codi-cmd-test-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`);
@@ -37,85 +28,6 @@ function cleanupTempDir(dir: string): void {
     fs.rmSync(dir, { recursive: true, force: true });
   } catch {
     // Ignore cleanup errors
-  }
-}
-
-/**
- * Process harness for E2E tests without PTY.
- */
-class ProcessHarness {
-  private proc: ChildProcess;
-  private output = '';
-  private exitPromise: Promise<number | null>;
-
-  constructor(command: string, args: string[], opts?: { cwd?: string; env?: NodeJS.ProcessEnv }) {
-    this.proc = spawn(command, args, {
-      cwd: opts?.cwd,
-      env: {
-        ...process.env,
-        ...opts?.env,
-        NO_COLOR: '1',
-        FORCE_COLOR: '0',
-        CI: '1',
-      },
-      stdio: ['pipe', 'pipe', 'pipe'],
-    });
-
-    this.proc.stdout?.on('data', (data) => {
-      this.output += data.toString();
-    });
-
-    this.proc.stderr?.on('data', (data) => {
-      this.output += data.toString();
-    });
-
-    this.exitPromise = new Promise((resolve) => {
-      this.proc.on('exit', (code) => resolve(code));
-      this.proc.on('error', () => resolve(null));
-    });
-  }
-
-  write(data: string): void {
-    this.proc.stdin?.write(data);
-  }
-
-  writeLine(data: string): void {
-    this.write(data + '\n');
-  }
-
-  getOutput(): string {
-    return this.output;
-  }
-
-  async waitFor(pattern: string | RegExp, timeoutMs = WAIT_TIMEOUT): Promise<string> {
-    const re = typeof pattern === 'string'
-      ? new RegExp(pattern.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'))
-      : pattern;
-
-    const start = Date.now();
-    while (Date.now() - start < timeoutMs) {
-      if (re.test(this.output)) {
-        return this.output;
-      }
-      await new Promise(r => setTimeout(r, 50));
-    }
-
-    throw new Error(`Timeout waiting for pattern: ${pattern}\n\nOutput:\n${this.output}`);
-  }
-
-  kill(): void {
-    this.proc.kill('SIGTERM');
-  }
-
-  async waitForExit(timeoutMs = 5000): Promise<number | null> {
-    const timeout = new Promise<number | null>((resolve) => {
-      setTimeout(() => {
-        this.kill();
-        resolve(null);
-      }, timeoutMs);
-    });
-
-    return Promise.race([this.exitPromise, timeout]);
   }
 }
